refactoring multiples activities into fragment and viewmodel
