+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,10 @@
         android:theme="@style/Theme.Factotum"
         tools:targetApi="31">
         <activity
-            android:name=".BasicActivity"
+            android:name=".maps.BasicActivity"
             android:exported="false"
             android:label="@string/title_activity_basic"
             android:theme="@style/Theme.Factotum.NoActionBar" />
-        <activity
-            android:name=".RoutesActivity"
-            android:exported="false" />
 
         <meta-data
             android:name="com.google.android.gms.version"
@@ -30,16 +27,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name=".MapsMarkerActivity"
-            android:exported="true"
-            android:label="@string/title_activity_maps">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
