+++ app/src/main/java/com/github/factotum_sdp/factotum/maps/SecondFragment.kt
@@ -1,8 +1,6 @@
-package com.github.factotum_sdp.factotum
+package com.github.factotum_sdp.factotum.maps
 
-import android.R
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
