--- app/src/main/res/layout/activity_routes.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/routesView"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".RoutesActivity">
-
-
-    <LinearLayout
-        android:id="@+id/routesLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical" >
-
-        <Button
-            android:id="@+id/startRoute"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="showRoute" />
-    </LinearLayout>
-</ScrollView>
