+++ app/src/main/res/layout/fragment_first.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FirstFragment">
+    tools:context=".maps.FirstFragment">
 
 
     <Button
